added ROR instruction
Former-commit-id: 1807af87f9d24d59741192aff4f5b827fe9e6bde

+++ src/util.h
@@ -19,8 +19,8 @@ void error  (std::string message);
 std::string to_hex_string(uint32_t val);
 
 // get nth bits from value as so: [start, end)
-inline uint16_t get_nth_bits(uint16_t val, uint8_t start, uint8_t end) {
-    return (val >> start) & (uint16_t)(pow(2, end - start) - 1);
+inline uint32_t get_nth_bits(uint32_t val, uint8_t start, uint8_t end) {
+    return (val >> start) & (uint32_t)(pow(2, end - start) - 1);
 }
 
 // get nth bit from value
